requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
face-recognition==1.3.0
pymongo==4.3.3
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
scipy==1.10.1
protobuf==4.22.3
